+++ src/checkers/cwe_190.mli
@@ -5,4 +5,4 @@ https://cwe.mitre.org/data/definitions/190.html
 val name : string
 val version : string  
 
-val check_cwe : Bap.Std.program Bap.Std.term -> Bap.Std.project -> Bap.Std.word Bap.Std.Tid.Map.t -> string list ->  unit
+val check_cwe : Bap.Std.program Bap.Std.term -> Bap.Std.project -> Bap.Std.word Bap.Std.Tid.Map.t -> string list list ->  unit