+++ src/checkers/cwe_248.mli
@@ -8,4 +8,4 @@ can actually catch the thrown exceptions, thus we generate some false negatives.
 val name : string
 val version : string
 
-val check_cwe : Bap.Std.program Bap.Std.term -> Bap.Std.word Bap.Std.Tid.Map.t -> unit
+val check_cwe : Bap.Std.program Bap.Std.term -> Bap.Std.project -> Bap.Std.word Bap.Std.Tid.Map.t -> string list list -> unit