+++ src/checkers/cwe_426.ml
@@ -22,4 +22,7 @@ let handle_sub sub program tid_map symbols =
   else ()
     
 let check_cwe program proj tid_map symbols =
-  Seq.iter (Term.enum sub_t program) ~f:(fun s -> handle_sub s program tid_map symbols)
+  match symbols with
+  | hd::[] ->
+     Seq.iter (Term.enum sub_t program) ~f:(fun s -> handle_sub s program tid_map hd)
+  | _ -> failwith "[CWE426] symbol_names not as expected"