+++ src/utils/json_utils.mli
@@ -1,0 +1,4 @@
+(* This module implements functionality related to parsing the JSON configuration file. *)
+
+val get_symbol_lists_from_json : Yojson.Basic.json -> string -> string list list
+val get_symbols_from_json : Yojson.Basic.json -> string -> string list