+++ CHANGES.md
@@ -1,3 +1,9 @@
+0.2-dev (2019-XX-XX)
+=====
+
+- Refactoring: Unification of cwe_checker function interface
+- Refactoring: Created utils module for JSON functionality
+
 0.1 (2018-10-08)
 =====